fix massive logging bug

+++ src/bmp.h
@@ -3,4 +3,4 @@
 void BMPStartMeasurement();
 void bmp_init(SPIClass* spi_to_use);
 void getBMPData(float *temperature, float *pressure);
-float altitude_from_pressure(uint32_t pressure);
+float altitude_from_pressure(float pressure);
